Reverted to previously committed files
Accidently commited code I didn't want to commit yet.

+++ units/units.go
@@ -1,10 +1,10 @@
 package units
 
 //1 kilobyte (1024 bytes)
-var KB int = 1024
+var KB int16 = 1024
 
 //1 megabyte (1048576 bytes)
-var MB int = 1048576
+var MB int32 = 1048576
 
 //1 gigabyte (1073741824 bytes)
-var GB int = 1073741824
+var GB int32 = 1073741824
